Updated entities

+++ cart-service/src/main/java/com/cg/cartservice/entities/OrderMain.java
@@ -53,7 +53,7 @@ public class OrderMain {
     private String buyerCity;
     private String buyerPincode;
     private BigDecimal orderAmount;
-    private Long discountedAmount;
+    private BigDecimal discountedAmount;
 
     @Enumerated(EnumType.STRING)
     private OrderStatus orderStatus;
