+++ cart-service/src/main/java/com/cg/cartservice/entities/ProductInOrder.java
@@ -57,6 +57,6 @@ public class ProductInOrder {
     private String productCategory;
     private BigDecimal productPrice;
     private Integer productStock;
-    private Long discountPercent;
+    private Double discountPercent;
 
 }
